+++ src/expansions/var_expansor.c
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/27 19:09:24 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/01/21 23:30:24 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/01/23 23:38:23 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,11 @@
 
 /**
  * * Get next index of a str to expand
- * @param str		str to expand it content
- * @return			str expanded
+ * @param str			str to expand it content
+ * @param is_heredoc	is called from heredoc
+ * @return				str expanded
 */
-static size_t	get_start_expand(const char *str)
+static size_t	get_start_expand(const char *str, int is_heredoc)
 {
 	int		quotes;
 	char	*start_expand;
@@ -29,9 +30,9 @@ static size_t	get_start_expand(const char *str)
 	{
 		if (quotes == NONE && *start_expand == '$')
 			return (start_expand - str);
-		else if (quotes == NONE && *start_expand == '\'')
+		else if (!is_heredoc && quotes == NONE && *start_expand == '\'')
 			quotes = SINGLE;
-		else if (quotes == SINGLE && *start_expand == '\'')
+		else if (!is_heredoc && quotes == SINGLE && *start_expand == '\'')
 			quotes = NONE;
 		start_expand++;
 	}
@@ -64,7 +65,7 @@ static char	*expand_status_at(char *str, size_t start_expand)
  * @param len_expand	lenght to replace
  * @return				str expanded
 */
-static char	*expand_at_free(char *malloc_str, t_str **env_list, \
+static char	*expand_at_free(char *malloc_str,  \
 				size_t start_expand, size_t len_expand)
 {
 	char	*oldset;
@@ -76,7 +77,7 @@ static char	*expand_at_free(char *malloc_str, t_str **env_list, \
 		out = expand_status_at(malloc_str, start_expand);
 	else
 	{
-		newset = ft_getenv(env_list, &oldset[1]);
+		newset = ft_getenv(&oldset[1]);
 		if (!newset)
 			out = ft_strreplaceat(malloc_str, oldset, "", start_expand);
 		else
@@ -91,16 +92,16 @@ static char	*expand_at_free(char *malloc_str, t_str **env_list, \
  * * Given str expand env variables ($ followed by characters) to their values
  * * Input str should come malloc
  * @param malloc_str	str to expand it content
- * @param env_list		environment list
+ * @param is_heredoc	is called from heredoc
  * @return				str expanded
 */
-char	*recursive_expand(char *malloc_str, t_str **env_list)
+char	*recursive_expand(char *malloc_str, int is_heredoc)
 {
 	size_t	start_expand;
 	size_t	len_expand;
 	char	*aux;
 
-	start_expand = get_start_expand(malloc_str);
+	start_expand = get_start_expand(malloc_str, is_heredoc);
 	if (start_expand != 0 || malloc_str[0] == '$')
 	{
 		len_expand = 1;
@@ -109,19 +110,19 @@ char	*recursive_expand(char *malloc_str, t_str **env_list)
 			len_expand++;
 		if (len_expand == 1 && aux[len_expand] == '?')
 			len_expand = 2;
-		malloc_str = expand_at_free(malloc_str, env_list, \
+		malloc_str = expand_at_free(malloc_str, \
 								start_expand, len_expand);
-		return (recursive_expand(malloc_str, env_list));
+		return (recursive_expand(malloc_str, is_heredoc));
 	}
 	return (malloc_str);
 }
 
 /**
  * * Given str expand env variables ($ followed by characters) to their values
- * @param str	str to expand it content
- * @return		str expanded
+ * @param str			str to expand it content
+ * @return				str expanded
 */
 char	*ft_expand(const char *str)
 {
-	return (recursive_expand(ft_strdup(str), &g_var.env));
+	return (recursive_expand(ft_strdup(str), FALSE));
 }